Remove Firebase support for Android and use QtFirebase instead
This is the first commit of many to come. I'm in a hurry so I'm not going in too much detail to fine tune the implementation.
A lot of things will change here. iOS support with QtFirebase will also be added. I'm taking this approach because I want qutils to be a library that you can use on all platforms with the same functionalities.
I'll add push notification support for macOS and Windows later and that's why I want to keep the `NotificationClient` class.